requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
openai==1.3.0
alpha-vantage==2.3.1
pandas==2.1.3
requests==2.31.0
python-dateutil==2.8.2
eventlet==0.33.3 